refactored to CSS modules, made sure max width is 100vw
