+++ src/css/Header.module.css
@@ -1,4 +1,4 @@
-header { 
+.header { 
   display: flex;
   flex-direction: column;
   align-items: center;
