+++ src/css/SearchBar.module.css
@@ -1,0 +1,3 @@
+.searchBar {
+  padding: 0px 2vw 10px 2vw;
+}
