--- src/css/SplashContainer.css
@@ -1,3 +1,0 @@
-#splash-button-container {
-  text-align: right;
-}
