+++ src/css/SplashContainer.module.css
@@ -1,0 +1,7 @@
+.splashContainer {
+  max-width: 100vw;
+}
+
+.splashButtonContainer {
+  text-align: right;
+}
