+++ src/css/SplashText.module.css
@@ -1,9 +1,9 @@
-#splash-text {
+.splashText {
   padding-left: clamp(5vw, 10vw, 20vw);
   text-align: right;
   font-size: clamp(1em, 3em, 4em);
 }
 
-#splash-text > span {
+#splashText > span {
   background: #A7BBCA;
 }
