+++ src/views/components/multi/Header.tsx
@@ -1,9 +1,9 @@
 import Logo from './Logo';
-import '../../../css/Header.css';
+import styles from '../../../css/Header.module.css';
 
 const Header = () => {
   return (
-    <header>
+    <header className={styles.header}>
       <Logo />
     </header>
   )
