+++ src/views/components/splash/SplashImage.tsx
@@ -1,10 +1,10 @@
 import splashImage from '../../../images/argent-splash-image.jpg';
-import '../../../css/SplashImage.css';
+import styles from '../../../css/SplashImage.module.css';
 
 const SplashImage = () => {
   return (
     <img
-      id="splash-image"
+      className={styles.splashImage}
       src={splashImage} 
       alt="An abstract design, shifting shapes of blue, purple, white, and black in a quartzlike/iridescent pattern" />
   )
