+++ src/views/components/splash/SplashText.tsx
@@ -1,8 +1,8 @@
-import '../../../css/SplashText.css';
+import styles from '../../../css/SplashText.module.css';
 
 const SplashText = () => {
   return (
-    <h2 id="splash-text">
+    <h2 className={styles.splashText}>
       Companies you've always <span>loved</span>. 
       Stocks you didn't know you <span>needed</span>. 
     </h2>
